Upgrade to work with guzzle6
